+++ Entities/BoardUser.cs
@@ -5,8 +5,8 @@ namespace HeroServer
 {
     public class BoardUser
     {
-        public int Id { get; set; }
-        public int WebSysUserId { get; set; }
+        public long Id { get; set; }
+        public long WebSysUserId { get; set; }
         public String FirstName1 { get; set; }
         public String FirstName2 { get; set; }
         public String LastName1 { get; set; }
@@ -20,7 +20,7 @@ namespace HeroServer
         {
         }
 
-        public BoardUser(int id, int webSysUserId, String firstName1, String firstName2, String lastName1, String lastName2, DateTime createDateTime, DateTime updateDateTime, int boardUserStatusId)
+        public BoardUser(long id, long webSysUserId, String firstName1, String firstName2, String lastName1, String lastName2, DateTime createDateTime, DateTime updateDateTime, int boardUserStatusId)
         {
             Id = id;
             WebSysUserId = webSysUserId;
@@ -33,7 +33,7 @@ namespace HeroServer
             BoardUserStatusId = boardUserStatusId;
         }
 
-        public BoardUser(int id, int webSysUserId, String firstName1, String firstName2, String lastName1, String lastName2, int boardUserStatusId)
+        public BoardUser(long id, long webSysUserId, String firstName1, String firstName2, String lastName1, String lastName2, int boardUserStatusId)
         {
             Id = id;
             WebSysUserId = webSysUserId;
