+++ Entities/Notification.cs
@@ -5,8 +5,8 @@ namespace HeroServer
 {
     public class Notification
     {
-        public int Id { get; set; }
-        public int WebSysUserId { get; set; }
+        public long Id { get; set; }
+        public long WebSysUserId { get; set; }
         public String MessageId { get; set; }
         public String Title { get; set; }
         public String Body { get; set; }
@@ -22,7 +22,7 @@ namespace HeroServer
         {
         }
 
-        public Notification(int id, int webSysUserId, String messageId, String title, String body, String action, String information, String parameter, int displayMode, DateTime dateTime, int notificationStatusId)
+        public Notification(long id, long webSysUserId, String messageId, String title, String body, String action, String information, String parameter, int displayMode, DateTime dateTime, int notificationStatusId)
         {
             Id = id;
             WebSysUserId = webSysUserId;
