+++ Functions/Access/PasswordRequest.cs
@@ -4,7 +4,7 @@ namespace HeroServer
 {
     public class PasswordRequest
     {
-        public int AppUserId { get; set; }
+        public long AppUserId { get; set; }
         public String AuthUserId { get; set; }
         public String NewPassword { get; set; }
 
@@ -12,7 +12,7 @@ namespace HeroServer
         {
         }
 
-        public PasswordRequest(int appUserId, String authUserId, String newPassword)
+        public PasswordRequest(long appUserId, String authUserId, String newPassword)
         {
             AppUserId = appUserId;
             AuthUserId = authUserId;
